Extract the progressive search engine into a reusable composable
